Move library update response class to it's own file
