+++ site/app/libraries/homework/UseCases/LibraryUpdateUseCase.php
@@ -5,34 +5,6 @@ use app\libraries\homework\Entities\LibraryEntity;
 use app\libraries\homework\Gateways\LibraryGateway;
 use app\libraries\homework\Gateways\Library\LibraryGatewayFactory;
 
-class LibraryUpdateResponse {
-    /** @var string */
-    public $error;
-
-    protected $message;
-
-    public function __construct(string $message) {
-        $this->message = $message;
-    }
-
-    /**
-     * @param string $message
-     * @return static
-     */
-    public static function error(string $message) {
-        $instance = new static('');
-        $instance->error = $message;
-        return $instance;
-    }
-
-    /**
-     * @return string
-     */
-    public function getMessage() {
-        return $this->message;
-    }
-}
-
 
 class LibraryUpdateUseCase extends BaseUseCase {
 
@@ -45,9 +17,9 @@ class LibraryUpdateUseCase extends BaseUseCase {
         $this->gateway = LibraryGatewayFactory::getInstance();
     }
 
-    public function updateLibrary($name): LibraryUpdateResponse {
+    public function updateLibrary($name): \app\libraries\homework\Responses\LibraryUpdateResponse {
         if (!$name) {
-            return LibraryUpdateResponse::error('You must specify the library to remove.');
+            return \app\libraries\homework\Responses\LibraryUpdateResponse::error('You must specify the library to remove.');
         }
 
         $library = new LibraryEntity($name, $this->location);
@@ -55,10 +27,10 @@ class LibraryUpdateUseCase extends BaseUseCase {
         $response = $this->gateway->updateLibrary($library);
 
         if ($response->success) {
-            return new LibraryUpdateResponse($response->message);
+            return new \app\libraries\homework\Responses\LibraryUpdateResponse($response->message);
         }
 
-        return LibraryUpdateResponse::error($response->message);
+        return \app\libraries\homework\Responses\LibraryUpdateResponse::error($response->message);
     }
 
 }
